Implement MEGA_FIX_PLAN_5 recovery-path hardening
H1 (hedge double-submit): add Hyperliquid orderStatus-by-cloid query and
gate cloid escalation behind it. The retry classifier now mints a new cloid
only after orderStatus proves the current cloid produced no fill, so a
WS-timeout-but-filled hedge can no longer be re-submitted into a double fill.
Replaces the dead confirm_recent_hyperliquid_fill (always returned None).

M1 (supervisor): add spawn_supervised_with_factory (bounded-backoff restart,
transient ServiceDown cleared on recovery, crash-loop latch) for the
reconstructable services; rename the old behavior to spawn_supervised_fail_closed
for the Receiver-owning services (cancel_manager, post_trade_auditor,
fill_detection).

M2/M3/L1 (reconciler): total_pending_qty + reconciler_grace_ms so it no longer
races an in-flight hedge; max_unhedged_hard_ms with progress-based timer reset
so a slow-but-working hedge isn't hard-errored; effective_dust = max(dust,
min_size) so sub-min unhedgeable residuals are tolerated instead of halting.

L2: single account_address() helper for HL wallet resolution across all services.
L3: position monitor reports cumulative (not per-tick) fills. L4: u128 dedup
quantizer. N4: clear stale unresolved_exposure.jsonl when flat. N5: config
validate guards (profit_rate>0, fees>=0). N6: drain idle hedge-WS reads.
N1/N2/N3: README credentials, 1kHz cadence, bounded-queue doc fixes.

cargo build --all-targets and cargo test (83 lib tests + suites) pass.

+++ src/services/fill_dedup.rs
@@ -12,12 +12,12 @@ pub enum FillKey {
     OrderId(u64),
     OrderCumulative {
         order_id: u64,
-        cumulative_quanta: u64,
+        cumulative_quanta: u128,
     },
     Cloid(String),
     CloidCumulative {
         cloid: String,
-        cumulative_quanta: u64,
+        cumulative_quanta: u128,
     },
 }
 
@@ -46,8 +46,11 @@ impl FillKey {
 }
 
 #[inline]
-fn quantize_size(size: f64) -> u64 {
-    (size.max(0.0) * 1_000_000_000.0).round() as u64
+fn quantize_size(size: f64) -> u128 {
+    // u128 so even ultra-cheap, very-large-size symbols (where size * 1e9 would
+    // overflow u64 and saturate to u64::MAX) cannot collapse distinct cumulative
+    // sizes onto the same dedup key.
+    (size.max(0.0) * 1_000_000_000.0).round() as u128
 }
 
 /// Bounded fill-dedup set.
